[Dockerfile][CMake] Update Dockerfile to correctly export acados variables. Automatically run python script to generate neural solver. Remove shortcuts for make process. Hotfix compilation error for wrench_est_actuator_meas_base.h

+++ aerial_robot_control/include/aerial_robot_control/wrench_est/wrench_est_actuator_meas_base.h
@@ -333,7 +333,7 @@ private:
       }
     }
 
-    ROS_INFO("Wrench Estimator: Entering state %d", state_);
+    ROS_INFO("Wrench Estimator: Entering state %d", static_cast<int>(state_));
     ros::NodeHandle wrench_est_nh(nh_, "controller/wrench_est");
     wrench_est_nh.setParam("state", static_cast<int>(state_));
   }
